backend/requirements.txt
# FastAPI 核心依赖
fastapi>=0.127.0
uvicorn[standard]>=0.40.0
python-multipart>=0.0.9

# 数据库
sqlalchemy>=2.0.0

# AI 模型
openai-whisper>=20231117

# 测试依赖
pytest>=9.0.0
httpx>=0.28.0
pytest-asyncio>=0.24.0

